+++ test/fixtures/changelog.md
@@ -1,12 +1,22 @@
-Change Log
+# Change Log
 
-All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.1.2 - Unreleased
 
 ## 0.1.1 - 2017-06-20
+
+
 ### Added
 - New feature
+- More things
+
+
+
+
 
 ### Fixed
 - Bug fix